refactor(prospect): extract the score + QR-route stages (P3-1 follow-up)
P3-1 lifted PERSIST and DISPATCH out of createProspect and left the earlier
stages inline, noting them as worth splitting when someone next had reason to
open the file. This takes the two cleanest remaining ones, verbatim:

  SCORE → prospectScoring.js    quiz re-scoring + profile-question validation
  ROUTE → prospectQrRouting.js  QR-level refinement of the internal path

createProspect drops ~780 → ~600 lines; prospectService.js 1,307 → 1,187.

Two closure side effects became explicit, which is most of the value:

  SCORE used to reach out and mutate incoming.sourceMetadata twice. It now
  returns a patch the caller merges, so what the stage writes is visible at the
  call site rather than happening behind it.

  ROUTE used to read AND reassign assignedAgentId/routeVia through the closure.
  They are an input and a returned output now, so it is legible that this stage
  can change who the lead belongs to.

Behaviour is unchanged. The externalAllow guard on every QR branch is preserved
exactly — external-eligible leads were already routed by resolveLeadAssignment,
which includes the QR tier, so re-running it here would double-route them.

Zero test edits: 9 prospect/quiz/capture suites (187 tests), unit 131/131
(2,222 tests), full integration tier 149/149 (2,614 tests), eslint clean.
